Changing deprecated implementation
